[Modify, Refactor] Separate review details Component and Apply isBestReview API
